NP-3522: move ViewingScope in its own file for testing
