+++ search-commons/src/main/java/no/unit/nva/search/restclients/responses/UserResponse.java
@@ -1,14 +1,8 @@
 package no.unit.nva.search.restclients.responses;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-
-import java.net.URI;
-import java.util.Collections;
-import java.util.Set;
-import org.elasticsearch.common.recycler.Recycler.V;
-
 import static no.unit.nva.search.constants.ApplicationConstants.objectMapperNoEmpty;
 import static nva.commons.core.attempt.Try.attempt;
+import com.fasterxml.jackson.core.JsonProcessingException;
 
 public class UserResponse {
 
@@ -29,43 +23,4 @@ public class UserResponse {
     public String toJson() {
         return attempt(() -> objectMapperNoEmpty.writeValueAsString(this)).orElseThrow();
     }
-
-    public static class ViewingScope {
-
-        private Set<URI> includedUnits;
-        private Set<URI> excludedUnits;
-        private boolean recursive;
-
-        public static ViewingScope create(URI... includedUnits){
-            var viewingScope = new ViewingScope();
-            viewingScope.setIncludedUnits(Set.of(includedUnits));
-            return viewingScope;
-        }
-
-        public Set<URI> getIncludedUnits() {
-            return includedUnits != null ? includedUnits : Collections.emptySet();
-        }
-
-        public void setIncludedUnits(Set<URI> includedUnits) {
-            this.includedUnits = includedUnits;
-        }
-
-        public Set<URI> getExcludedUnits() {
-            return excludedUnits != null ? excludedUnits : Collections.emptySet();
-
-        }
-
-        public void setExcludedUnits(Set<URI> excludedUnits) {
-            this.excludedUnits = excludedUnits;
-        }
-
-        public boolean isRecursive() {
-            return recursive;
-        }
-
-        public void setRecursive(boolean recursive) {
-            this.recursive = recursive;
-        }
-    }
-
 }
