+++ search-commons/src/test/java/no/unit/nva/search/SearchClientTest.java
@@ -3,6 +3,7 @@ package no.unit.nva.search;
 import no.unit.nva.search.models.Query;
 import no.unit.nva.search.models.SearchResourcesResponse;
 import no.unit.nva.search.restclients.responses.UserResponse;
+import no.unit.nva.search.restclients.responses.ViewingScope;
 import nva.commons.apigateway.exceptions.ApiGatewayException;
 import nva.commons.apigateway.exceptions.BadGatewayException;
 import nva.commons.core.paths.UriWrapper;
@@ -78,8 +79,8 @@ public class SearchClientTest {
         assertNotNull(response);
     }
 
-    private UserResponse.ViewingScope getSampleViewingScope() {
-        UserResponse.ViewingScope viewingScope = new UserResponse.ViewingScope();
+    private ViewingScope getSampleViewingScope() {
+        ViewingScope viewingScope = new ViewingScope();
         viewingScope.setIncludedUnits(Set.of(randomUri(), randomUri()));
         viewingScope.setExcludedUnits(Set.of(randomUri()));
         return viewingScope;
